refactor PlotSurface, add axis value to MapSlider text
