+++ app/src/main/java/com/kodeco/android/coordplot/app/widgets/Map.kt
@@ -1,4 +1,5 @@
 package com.kodeco.android.coordplot.app.widgets
+import android.content.res.Resources.Theme
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
