+++ app/src/main/java/com/kodeco/android/coordplot/app/widgets/MapSlider.kt
@@ -5,6 +5,7 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Slider
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
@@ -14,12 +15,13 @@ import com.kodeco.android.coordplot.ui.theme.MyApplicationTheme
 
 @Composable
 fun MapSlider(titleRes: Int, percentage: Float, onPercentageChange: (Float) -> Unit) {
-    Row {
+    Row(verticalAlignment = Alignment.CenterVertically) {
         Text(
-            text = stringResource(id = titleRes),
+            text = "${stringResource(id = titleRes)}: ${String.format("%.0f", percentage * 100)}%",
             modifier = Modifier
                 .padding(start = 8.dp)
-                .width(120.dp))
+                .width(120.dp)
+        )
         Slider(value = percentage, onValueChange = onPercentageChange)
     }
